requirements.txt
streamlit>=1.25
ollama>=0.0.0
python-dotenv>=1.0.0
# Note: 'ollama' package name assumed from docs. If installation differs, adjust accordingly.
